requirements.txt
amqp==5.1.1
async-timeout==4.0.2
bidict==0.22.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==1.6.0
Deprecated==1.2.13
Flask==2.2.2
Flask-SocketIO==5.3.1
gearsclient==1.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
python-engineio==4.3.4
python-socketio==5.7.2
redis==3.5.3
requests==2.28.1
simple-websocket==0.8.1
six==1.16.0
urllib3==1.26.12
vine==5.0.0
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
